feat: refactor code and add new zoom icon
