+++ src/constants/brushTools.jsx
@@ -6,5 +6,6 @@ export const polygon = require('./../assets/Images/Tools/polygon.png');
 export const polyline = require('./../assets/Images/Tools/polyline.png');
 export const clear = require('./../assets/Images/Tools/clear.png');
 export const remove = require('./../assets/Images/Tools/remove.png');
+export const zoom = require('./../assets/Images/Tools/zoom.png');
 
-export const brushTools = {brush, eraser, lasso, pan, polygon, polyline, clear, remove};
+export const brushTools = {brush, eraser, lasso, pan, polygon, polyline, clear, remove, zoom};
