+++ src/views/FileHandler/FileHandler.scss
@@ -12,7 +12,6 @@
 
   .button-container {
     display: flex;
-    justify-content: space-around;
     width: 100%;
 
     .file-name {
